+++ usaspending_api/recipient/models.py
@@ -167,3 +167,16 @@ class SummaryAwardRecipient(models.Model):
     class Meta:
         managed = True
         db_table = "summary_award_recipient"
+
+
+class RecipientAgency(models.Model):
+    id = models.IntegerField(primary_key=True)
+    fiscal_year = models.IntegerField(db_index=True)
+    toptier_code =  models.TextField(db_index=True)
+    recipient_hash = models.UUIDField(db_index=True)
+    recipient_name = models.TextField()
+    recipient_amount = models.DecimalField(max_digits=23, decimal_places=2)
+
+    class Meta:
+        managed = True
+        db_table = "recipient_agency"
